requirements.txt
requests
django
Pillow
psycopg2
djangorestframework
phonenumbers
redis
https://github.com/gazoon/fakeredis/archive/master.zip
django-bulk-update
pika
